Create Day-5: Password_Generator.py
This program is a password generator that creates a random, secure password based on user preferences for the number of letters, symbols, and numbers. It uses Python’s random module to select characters randomly from predefined lists and combines them into a shuffled sequence for unpredictability. Finally, it joins the characters into a string and displays the generated password to the user.